week2 문제B 완성 v2.0 / 잘못된 입력값만 재입력 할 수 있도록 수정 완료
